[step3] feat: Controller 추가 및 로또 버전별 실행 기능 추가
